add react-router and new components for creating a new event and viewing an exiting event
