+++ client/src/App.js
@@ -1,63 +1,17 @@
 import React, { Component } from 'react';
-import EventContract from './contracts/Event.json'
-import getWeb3 from './utils/getWeb3'
+import { BrowserRouter, Route } from 'react-router-dom'
+import NewEvent from './NewEvent'
+import Event from './Event'
 
 class App extends Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      web3: null,
-      tickets: []
-    }
-  }
-
-  componentWillMount() {
-    getWeb3
-    .then(results => {
-      this.setState({ web3: results.web3 })
-      this.instantiateContract()
-    })
-    .catch(() => {
-      console.log('Error finding web3.')
-    })
-  }
-
-  instantiateContract() {
-    const contract = require('truffle-contract')
-    const eventContract = contract(EventContract)
-    eventContract.setProvider(this.state.web3.currentProvider)
-
-    eventContract.deployed().then((instance) => {
-      instance.getTickets.call().then((data) => {
-        let tickets = []
-        let numTickets = data[0].length
-        for (let i = 0; i < numTickets; i++) {
-          tickets.push({
-            identifier: data[0][i].toNumber(),
-            price: data[1][i].toNumber()
-          })
-        }
-        return this.setState({tickets: tickets})
-      })
-    })
-  }
-
-  render() {
+  render () {
     return (
-      <div>
-        <ul>
-          {this.state.tickets.map(this.renderTicket)}
-        </ul>
-      </div>
-    );
-  }
-
-  renderTicket(ticket) {
-    return (
-      <li key={ticket.identifier}>
-        {ticket.identifier} {ticket.price}
-      </li>
+      <BrowserRouter>
+        <div>
+          <Route exact path="/" component={NewEvent}/>
+          <Route path="/:contractAddress" component={Event}/>
+        </div>
+      </BrowserRouter>
     )
   }
 }
