+++ client/src/NewEvent.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class NewEvent extends Component {
+  render () {
+    return (
+      <div>
+        New Event
+      </div>
+    )
+  }
+}
+
+export default NewEvent
